bugtrap updated to the latest version
(cherry picked from commit 15de7be5783243480ee146cf42294c8537c5aca9)

+++ code/3rd-party/bugtrap/CMapi.h
@@ -120,7 +120,7 @@ private:
 	/// Initialize particular recipient record.
 	void InitRecipient(ULONG ulRecipClass, MapiRecipDesc& rRecipDesc, const CStrHolder& strName);
 	/// Initialize recipients list.
-	void InitRecipients(ULONG ulRecipClass, lpMapiRecipDesc lpRecips, int& nRecipIndex, const CArray<CStrHolder>& arrRecipients);
+	void InitRecipients(ULONG ulRecipClass, lpMapiRecipDesc lpRecips, size_t& nRecipIndex, const CArray<CStrHolder>& arrRecipients);
 
 	/// Mapi session handle.
 	LHANDLE m_hSession;
